+++ packages/xlucene-evaluator/src/interfaces.ts
@@ -16,3 +16,25 @@ export interface GeoDistanceObj {
     distance: number;
     unit: GeoDistanceUnit;
 }
+
+export interface GeoPoint {
+    lat: number;
+    lon: number;
+}
+
+export type GeoDistanceUnit = 'miles'|'yards'|'feet'|'inch'|'kilometers'|'meters'|'centimeters'|'millimeters'|'nauticalmiles';
+
+export enum FieldType {
+    Geo = 'geo',
+    Date = 'date',
+    IP = 'ip',
+    String = 'string',
+    Integer = 'integer',
+    Float = 'float',
+    Boolean = 'boolean',
+    Object = 'object'
+}
+
+export interface TypeConfig {
+    [field: string]: FieldType;
+}
